Bullet changes
