+++ paranomal-game/Assets/Prefabs/5.56mm AP.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: -3794316438491666664}
   - component: {fileID: -8007417334808695360}
   - component: {fileID: -537775593009845427}
+  - component: {fileID: -5057849226855385570}
   m_Layer: 7
   m_Name: 5.56mm AP
   m_TagString: Untagged
@@ -128,12 +129,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 487c1a877d003fc45aa942f8a5c89036, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  damageDealt: 110
   addedRange: 17.17
   isArmourPiercing: 1
   isSoftPoint: 0
   bulletBody: {fileID: -3794316438491666664}
-  projectileWeight: 0.0034
+  projectileWeight: 0.34
 --- !u!96 &-537775593009845427
 TrailRenderer:
   serializedVersion: 2
@@ -242,3 +242,16 @@ TrailRenderer:
   m_MinVertexDistance: 0.1
   m_Autodestruct: 0
   m_Emitting: 1
+--- !u!114 &-5057849226855385570
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3930647487463510722}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4f7db67542659a64e8bb0502f91b964c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damage: 110
