+++ paranomal-game/Assets/Prefabs/9mm SP.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: 3484965901773759690}
   - component: {fileID: 7237962263235450439}
   - component: {fileID: 1849898791912401675}
+  - component: {fileID: -6781408416689383395}
   m_Layer: 7
   m_Name: 9mm SP
   m_TagString: Untagged
@@ -236,9 +237,21 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 487c1a877d003fc45aa942f8a5c89036, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  damageDealt: 100
   addedRange: 0
   isArmourPiercing: 0
   isSoftPoint: 1
   bulletBody: {fileID: 3484965901773759690}
-  projectileWeight: 0.00745
+  projectileWeight: 0.745
+--- !u!114 &-6781408416689383395
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8744822132129637132}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4f7db67542659a64e8bb0502f91b964c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damage: 100
