+++ paranomal-game/Assets/Scripts/Bullet/BulletDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f7db67542659a64e8bb0502f91b964c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
