+++ paranomal-game/Assets/Scripts/Events/EventOnlyInteractable.cs
@@ -1,0 +1,4 @@
+public class EventOnlyInteractable : Interactable
+{
+
+}
