+++ paranomal-game/Assets/Scripts/Events/EventOnlyInteractable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c94dff7acc1ff494bb76fb39be98a0e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
