--- paranomal-game/Assets/Scripts/InputManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ec9bad6c7365422438bb81798aced505
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
