+++ paranomal-game/Assets/Scripts/InputManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c8af2b3a21fe5374cbf5ced7fd00eda8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
