--- paranomal-game/Assets/Scripts/Interactable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 086ebc3ffb4ba914981482ba7ff8fa50
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
