+++ paranomal-game/Assets/Scripts/Interactables/InteractionEvent.cs
@@ -1,0 +1,7 @@
+using UnityEngine.Events;
+using UnityEngine;
+
+public class InteractionEvent : MonoBehaviour
+{
+    public UnityEvent OnInteract;
+}
