+++ paranomal-game/Assets/Scripts/Interactables/InteractionEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6125b64ec89ee294786136c2b133d6f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
